requirement.txt
# SL-PAM and PALM algorithm applied to joint image restauration and edges detection
 
# Requirements:

numpy==1.18.0
matplotlib==3.4.1
scipy==1.4.1

# pillow>=8.0.1
Pillow == 9.2.0

# progressbar
pip install progressbar2

# open-cv
pip install opencv-python

# skimage
python -m pip install -U scikit-image

